requirements.txt
# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Configuration management
pydantic==2.5.0
python-dotenv==1.0.0

# OCR related
paddlepaddle==2.5.2
paddleocr==2.7.0.3
Pillow==10.1.0

# PDF processing
PyMuPDF<1.21.0

# LLM API
openai>=1.12.0
tiktoken>=0.5.1

# Data processing
numpy==1.24.3
pandas==2.0.3

# Utilities
aiofiles==23.2.1
typing-extensions==4.8.0

# Logging and monitoring
loguru==0.7.2

# Development tools
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0 